+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 # Changelog
 
-## 1.6.1 (2022-12-23)
+## 1.8.0 (2022-12-24)
+
+### New Features
+
+- Implement `doc` command
+
+### Bug Fixes
+
+- Rename `settings` command to `command`
+
+## 1.7.1 (2022-12-23)
 
 ### Bug Fixes
 
 - Fix wording of commands
 
-## 1.6.0 (2022-12-23)
+## 1.7.0 (2022-12-23)
 
 ### New Features